EnemyWAVES AND AcclerationWaves

+++ Assets/Scripts/Imported Scripts/LevelResultController.cs
@@ -84,7 +84,7 @@ namespace SpaceShooter
             TotalStats.score += Player.Instance.Score;
 
             // бонус за время прохождения.
-            int timeBonus = LevelController.Instance.ReferenceTime - (int)LevelController.Instance.LevelTime;
+            int timeBonus = (int)( LevelController.Instance.ReferenceTime - (int)LevelController.Instance.LevelTime);
 
             if (timeBonus > 0)
                 TotalStats.score += timeBonus;
